Remove setting that requires higher API version
The result of setting distributed_routing works properly when
the operation fails. It doesn't work well when the operation succeeds,
as the current API version does not return the updated setting, but
later API versions do. Thus we postpone the activation of this
set to after we bump up the API version.