Change media item creation to be optional
media item creation was listed as mandatory in TESTING.md
(and consequently users needed to provide its name)
However, given that this item is not used in terraform-provider-vcd
testing but only in go-vcloud-director, it was moved to the optional
items. Leaving its name empty will skip the creation.
Also, add check to make sure all vital elements are filled.