Add exception fields to datasource test
The datasource test assumes that adding the mandatory tests is enough to
prove that the error handling is correct. However, some data sources
have an alternate way to using identification fields.
By adding a list of exceptions (optional fields that should be included
in the data source definition) we keep the test functioning and the data
sources elastic.