Add functions to find a catalog without its Org
These changes fix the problem of a catalog that was shared
to the current org, but a tenant could not retrieve it
for lack of privileges.
It also fixes the case when multiple catalogs with the same
name are available in the organization.
The new functions search for catalogs without need of accessing
the sharing organization.
Also added a new  'is_local' property to all catalog resources and data sources.
This property tells whether the catalog was created in the current org
or not.

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>